requirements.txt
python-socketio
nonebot2